Don't execute this query when it's going to fail.
While this doesn't look very important, it is; with PostgreSQL any failing SQL
stament done inside a transaction makes the entire transaction to rollback when
calling RecipeDB::commit().

+++ src/backends/qsqlrecipedb.cpp
@@ -841,9 +841,11 @@ void QSqlRecipeDB::saveRecipe( Recipe *recipe )
 	}
 
 	// only delete those ratings that don't exist anymore
-	command = QString( "DELETE FROM ratings WHERE recipe_id=%1 AND id NOT IN( %2 )" )
-	          .arg( recipeID ).arg( ids.join(",") );
-	recipeToSave.exec( command );
+	if ( !ids.isEmpty() ) {
+		command = QString( "DELETE FROM ratings WHERE recipe_id=%1 AND id NOT IN( %2 )" )
+			  .arg( recipeID ).arg( ids.join(",") );
+		recipeToSave.exec( command );
+	}
 
 	if ( newRecipe )
 		emit recipeCreated( Element( recipe->title.left( maxRecipeTitleLength() ), recipeID ), recipe->categoryList );
